Add first websocket test
Introduced an abstraction above socket.io so the test stack can
use its own socket.io server.
I'm not happy exposing the client socket, but it was the only way I
could release the resources in the afterEach block without triggering a
timeout. There is probably a better way of doing this.